Split the pdf_file from the Paper model to the Revision model
This change was done so that the interface for getting revisions
for a paper is much simpler (it is just a matter of getting the most
recent revision).

This change was complicated by an issue with removing fields from a
model that do no allow null values[0]. A fix was suggested on SO,
but for whatever reason it did not seem to work[1]. So the next best
fix was to just split the migration into two steps. The first step
creates changes the pdf_file field so that it can be null and it
does the actual migration to the Revision model. The second step
just goes through and deletes the pdf_file field from the Paper
model.

[0] https://docs.djangoproject.com/en/1.8/ref/migration-
operations/#removefield
[1] http://stackoverflow.com/questions/28878428/revert-django-1-7
-removefield-migration